Complete translation PL
Adding a new text string to the PL translation.

I propose to add text strings for translation in Transifex. You can do this with your team or submit a request to join the Nextcloud team. This will allow you to systematically translate the added text strings and translate into multiple languages 😉.